requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
openai==1.3.6
aiofiles==23.2.0
pydantic==2.5.0
sqlalchemy==2.0.23
email-validator==2.1.0